Enhance README with Logic Apps identity usage
The setup code/basic deploy/bicep only sets a single managedIdId to use the internal managed id. This means that in the Logic Apps UI, only the internal managed account is available in the dropdown when the HTTP Connector is set to Managed Identity. As such, to use the external managed ID the identity must be added to the Logic Apps that will be used to manage Marketplace subscriptions.